requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
xgboost>=1.5.0
torch>=1.10.0
transformers>=4.20.0
sentence-transformers>=2.2.0
scipy>=1.7.0
openai>=0.27.0
anthropic>=0.3.0
PyGithub>=1.55
joblib>=1.1.0
tqdm>=4.62.0
requests>=2.27.0
imbalanced-learn>=0.9.0
squarify>=0.4.3
psutil>=5.8.0
